backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
numpy
librosa
praat-parselmouth
google-generativeai
python-dotenv
gTTS
